New folder/Jarvis/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0

# LangGraph & LangChain
langgraph==0.0.51
langchain==0.1.0
langchain-core==0.1.2
langchain-anthropic==0.1.0

# AI APIs
anthropic==0.7.8
openai==1.3.8
pinecone-client==3.0.2

# Google (Gmail, OAuth)
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Financial (Plaid)
plaid-python==17.8.0

# DevOps & Deployment
PyGithub==2.1.1
requests==2.31.0

# Logging & Utils
winston==0.0.1
python-logging==0.4.9.7

# Async
asyncio==3.4.3
aiosqlite==3.14.0

# Redis (for production task queues)
redis==5.0.1
